use bigint

+++ frontend/api/internal_api/deleteInternalUser.ts
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client';
 
 export const deleteInternalUser = gql`
-  mutation DeleteInternalUser($id: Int!) {
+  mutation DeleteInternalUser($id: BigInt!) {
     deleteInternalUser(id: $id) {
       id
       email
